lsu-pilot/pyproject.toml
[tool.poetry]
name = "lsu-pilot"
version = "0.1.0"
description = ""
authors = ["Kody Buss <kody.buss@zelis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.6"
openai = "^1.52.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
tiktoken = "^0.8.0"
langchain = "^0.3.4"
numpy = "^1.21"
scipy = "^1.14.1"
cairosvg = "^2.7.1"
requests = "^2.32.3"
replicate = "^1.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
